requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
asyncio-mqtt>=0.13.0

# AI and ML
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0
tiktoken>=0.5.2

# HTTP and networking
aiohttp>=3.9.0
httpx>=0.25.0
websockets>=12.0

# Data handling
pandas>=2.1.0
# sqlite3 is built-in with Python
pyyaml>=6.0.1

# CLI and UI
rich>=13.6.0
typer>=0.9.0
click>=8.1.0

# Monitoring and observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
structlog>=23.2.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
pre-commit>=3.5.0

# Security
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Configuration
python-dotenv>=1.0.0
dynaconf>=3.2.0

# Database
alembic>=1.12.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0

# Caching
redis>=5.0.0
aioredis>=2.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0
mkdocs-mermaid2-plugin>=1.1.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
requests>=2.31.0
aiofiles>=23.2.0